Extract renderer session view reducer
